taifex_data_pipeline/requirements.txt
python-magic
patool
pytz # 用於時區處理
